Week 5 Lab Order Entry Phase 2
Put the wrong week on the other last commit, oh well.
I hadn't formatted any of the numbers (money, percents) and I really
didn't want to loose points for something that silly after all that work
with the tabs, so... might as well re-work how I do that anyway.
- Printing now runs over a list, so adding new stuff is as simple as
adding your input/calculation and then adding the label/value pair to
that list; no more variables per pair and super long println's.
-Added a function that preserves up to 2 decimal places while
eliminating trailing zeroes - adapted from others' work on
StackOverflow.
The fmt function was only required because 0.15f * 100 is something like
15.000001 and I don't like trailing zeroes.
At this point, this is something I'd be okay with shoving into a
production model. The only thing with room to expand is the stringTabber
function, as that could be reworked to except a list of n length and tab
all items within. That actually sounds like a lot of fun to make, but
not right now.
The most upsetting part about all of this is that the last lesson in my
java course was on operators. Going from talking about the difference
between division and modulus to figuring out how to dynamically
tab-align strings of unknown length is something else. Of course, I
could have (and should have) just manually tabbed my sample data - but
I'm too insane for something that intelligent.