ultimateStringTabber
Playing around with making a string tabber function that can handle
lists of n size. The function seems to work just find, though my
implementation in the test Main looses ordering of the columns. That
seems like a problem that would be fixed by using something other than a
dictionary - which makes it an implementation problem, not one with the
function.
Lots of random, unused code here because experimentation.