Week 6 Lab 2 - Powers Table
Simple program to display the square and cube of every integer up to the
user's input. Everything is placed into a single string for printing, as
per assignment instructions. Tab placement is manual, but thanks to the
lengths of the headers, that's not an issue until you need to display
the square of 3163. At that point, there will be so many rows that using
my ultimateStringTabber will probably take longer than the heat death of
the universe - never mind the console's ability to display that many
lines anyways.
It was pretty fun discovering the hard way that caret is not
exponentiation but is, in fact, bitwise OR. I got just a little confused
when I saw in an example on W3 Schools that 5 ^= 3 results in 6.