Week 7 - Order Entry Phase 3
Now this is pod racing
Did some experimentation already with expanding my stringTabber function
to accept a list of any length to tab; now, imagine my gleeful shrieking
when I saw that making multiple rows was the task of phase 3. Seriously
though, this code was pretty fun to write. Wrapping my head around how I
was going to store a list of n length of input for 9+ columns, tab all
of that evenly, and then print it by row was a delightful challenge.
Having to add so many things manually is a bit of a bummer, but I
suppose that is just the header row(s) and the total line. If this were
a production program, it would make adding new columns a bit of a
hassle, but I'm not sure at this time how to fix it - especially given
that this isn't a production program so it likely would not be worth the
effort.
Also gosh diggity dang it I lost a few points on phase 2 because the
professor calculates the discount on the gross + tax instead of just the
gross. It feels counterintuitive to provide a discount on the amount of
tax someone has to pay (and the IRS probably has opinions to share too),
but whatever. This is how it is in the sample output, and so too shall
it be in mine.

I also updated my Java/Eclipse IDE and seems to have changed some stuff
in older projects. I'm not sure what at this time, but perhaps I'll take
a peek post-commit.