Week 8 - Lab 2
Exceptionally basic program to take an input value and show the tip
amount and total at 3 preset rates. I know it is possible to use the
java.text.NumberFormat to format percentages, but at this point easier
to just copy over fmt and multiply by 100 myself.
I suppose we can't have every week be an epic struggle to violate the
laws of nature, physics, and morality via the Java programming language.