Week 9 - Course Project Phase 4 Lab
Working on the biggie again. This time, another table is added in - only
one row of input, so nothing major. The instructions wanted us to use a
single String variable to store *everything* but that's pretty goofy.
You could do that, sure, but not if you want automatic tabbing.
Hopefully, I don't loose points for not doing that. Phase 3 got a 100%,
so hopefully I'm good.
Implemented the java.text.NumberFormat class. No more manual formatting
of money or percents! Only downside is that percFormatter expects a
decimal percent (0.10 --> 10%), which does not align with the input
format - some processing is still required.
I have to wonder if there is a more performant way to handle this. Most
of what I've done has just been first idea type stuff; I haven't really
gone back and refactored for performance. I'm not really sure what I
could do to improve this, but it'd be interesting to see.