Week 10 - Vacation Diary Part 2
Combination data validation and multiple entries. Ended up settling for
using spaces to align the output as the assignment specifically required
making the city and country text exactly 15 characters - either via
adding spaces or truncating the input. Tabs are 8 characters in my
console, so everything looked pretty mediocre aligned by tabs. That
combined with the really hefty headers (also in the sample output) meant
it was easier to just hard-code it. Aside from that, standard classwork.